- Initial work converting serverspec to Inspec. - Temporarily disabled OSX and Windows integrations
