+++ .kitchen.docker.yml
@@ -10,6 +10,9 @@ transport:
 provisioner:
   name: dokken
 
+verifier:
+  name: inspec
+
 platforms:
 # - name: centos-5
 #   driver:
