+++ test/integration/helpers/support/omnibus_build.rb
